chatbot/requirements.txt
langchain
langchain-huggingface
streamlit
python-dotenv